+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 ### Added
 - Bug report and feature request issue templates
 - Code of conduct
+- CI/CD container
+- SemaphoreCI testing for Backend
 
 ### Changed
 - Done misc updates on README
@@ -14,7 +16,6 @@
 - Bumped Flask-Migrate from 2.2.1 to 2.3.1 in /frontend
 - Bumped WTForms from 2.1 to 2.2.1 in /frontend
 - Bumped SQLAlchemy from 1.2.5 to 1.2.16 in /frontend
-- Bumped PyYAML from 3.13 to 4.2b4 in /frontend
 - Bumped Flask-JWT-Extended from 3.13.1 to 3.15.0 in /frontend
 - Bumped pip from 10.0.1 to 18.1 in /frontend
 - Bumped PyYAML from 3.13 to 4.2b4 in /backend
@@ -34,6 +35,7 @@
 ### Security
 - Bumped flask from 0.12.2 to 1.0.2 in /frontend
 - Bumped requests from 2.19.1 to 2.21.0 in /frontend
+- Bumped PyYAML from 3.13 to 4.2b4 in /frontend
 
 ## [1.0] - 2018-12-20
 ### Added (initial Apollo release)